♻️ Adjust code structure
