+++ src/chatgpt.ts
@@ -1,7 +1,7 @@
 import { ChatGPTAPI, ChatGPTAPIBrowser } from "chatgpt";
-import { Message } from "wechaty";
+
 import { config } from "./config.js";
-import { ContactInterface, RoomInterface } from "wechaty/impls";
+
 import {
   IChatGPTItem,
   IConversationItem,
@@ -9,28 +9,6 @@ import {
   IAccount,
 } from "./interface.js";
 
-enum MessageType {
-  Unknown = 0,
-
-  Attachment = 1, // Attach(6),
-  Audio = 2, // Audio(1), Voice(34)
-  Contact = 3, // ShareCard(42)
-  ChatHistory = 4, // ChatHistory(19)
-  Emoticon = 5, // Sticker: Emoticon(15), Emoticon(47)
-  Image = 6, // Img(2), Image(3)
-  Text = 7, // Text(1)
-  Location = 8, // Location(48)
-  MiniProgram = 9, // MiniProgram(33)
-  GroupNote = 10, // GroupNote(53)
-  Transfer = 11, // Transfers(2000)
-  RedEnvelope = 12, // RedEnvelopes(2001)
-  Recalled = 13, // Recalled(10002)
-  Url = 14, // Url(5)
-  Video = 15, // Video(4), Video(43)
-  Post = 16, // Moment, Channel, Tweet, etc
-}
-
-const SINGLE_MESSAGE_MAX_SIZE = 500;
 const ErrorCode2Message: Record<string, string> = {
   "503":
     "OpenAI 服务器繁忙，请稍后再试| The OpenAI server is busy, please try again later",
@@ -38,10 +16,12 @@ const ErrorCode2Message: Record<string, string> = {
     "OpenAI 服务器限流，请稍后再试| The OpenAI server was limited, please try again later",
   "500":
     "OpenAI 服务器繁忙，请稍后再试| The OpenAI server is busy, please try again later",
+  "403":
+    "OpenAI 服务器拒绝访问，请稍后再试| The OpenAI server refused to access, please try again later",
   unknown: "未知错误，请看日志 | Error unknown, please see the log",
 };
 const Commands = ["/reset", "/help"] as const;
-export class ChatGPTPoole {
+export class ChatGPTPool {
   chatGPTPools: Array<IChatGPTItem> | [] = [];
   conversationsPool: Map<string, IConversationItem> = new Map();
   async resetAccount(account: IAccount) {
@@ -194,130 +174,5 @@ export class ChatGPTPoole {
       }
     }
     return ErrorCode2Message.unknown;
-  }
-}
-export class ChatGPTBot {
-  // Record talkid with conversation id
-  chatGPTPool = new ChatGPTPoole();
-  chatPrivateTiggerKeyword = config.chatPrivateTiggerKeyword;
-  botName: string = "";
-  ready = false;
-  setBotName(botName: string) {
-    this.botName = botName;
-  }
-  get chatGroupTiggerKeyword(): string {
-    return `@${this.botName}`;
-  }
-  async startGPTBot() {
-    console.debug(`Start GPT Bot Config is:${JSON.stringify(config)}`);
-    await this.chatGPTPool.startPools();
-    console.debug(`🤖️ Start GPT Bot Success, ready to handle message!`);
-    this.ready = true;
-  }
-  // TODO: Add reset conversation id and ping pong
-  async command(): Promise<void> {}
-  // remove more times conversation and mention
-  cleanMessage(rawText: string, privateChat: boolean = false): string {
-    let text = rawText;
-    const item = rawText.split("- - - - - - - - - - - - - - -");
-    if (item.length > 1) {
-      text = item[item.length - 1];
-    }
-    text = text.replace(
-      privateChat ? this.chatPrivateTiggerKeyword : this.chatGroupTiggerKeyword,
-      ""
-    );
-    // remove more text via - - - - - - - - - - - - - - -
-    return text;
-  }
-  async getGPTMessage(text: string, talkerId: string): Promise<string> {
-    return await this.chatGPTPool.sendMessage(text, talkerId);
-  }
-  // The message is segmented according to its size
-  async trySay(
-    talker: RoomInterface | ContactInterface,
-    mesasge: string
-  ): Promise<void> {
-    const messages: Array<string> = [];
-    let message = mesasge;
-    while (message.length > SINGLE_MESSAGE_MAX_SIZE) {
-      messages.push(message.slice(0, SINGLE_MESSAGE_MAX_SIZE));
-      message = message.slice(SINGLE_MESSAGE_MAX_SIZE);
-    }
-    messages.push(message);
-    for (const msg of messages) {
-      await talker.say(msg);
-    }
-  }
-  // Check whether the ChatGPT processing can be triggered
-  tiggerGPTMessage(text: string, privateChat: boolean = false): boolean {
-    const chatPrivateTiggerKeyword = this.chatPrivateTiggerKeyword;
-    let triggered = false;
-    if (privateChat) {
-      triggered = chatPrivateTiggerKeyword
-        ? text.includes(chatPrivateTiggerKeyword)
-        : true;
-    } else {
-      triggered = text.includes(this.chatGroupTiggerKeyword);
-    }
-    if (triggered) {
-      console.log(`🎯 Triggered ChatGPT: ${text}`);
-    }
-    return triggered;
-  }
-  // Filter out the message that does not need to be processed
-  isNonsense(
-    talker: ContactInterface,
-    messageType: MessageType,
-    text: string
-  ): boolean {
-    return (
-      talker.self() ||
-      messageType > MessageType.GroupNote ||
-      talker.name() == "微信团队" ||
-      // 语音(视频)消息
-      text.includes("收到一条视频/语音聊天消息，请在手机上查看") ||
-      // 红包消息
-      text.includes("收到红包，请在手机上查看") ||
-      // 位置消息
-      text.includes("/cgi-bin/mmwebwx-bin/webwxgetpubliclinkimg")
-    );
-  }
-
-  async onPrivateMessage(talker: ContactInterface, text: string) {
-    const talkerId = talker.id;
-    const gptMessage = await this.getGPTMessage(text, talkerId);
-    await this.trySay(talker, gptMessage);
-  }
-
-  async onGroupMessage(
-    talker: ContactInterface,
-    text: string,
-    room: RoomInterface
-  ) {
-    const talkerId = room.id + talker.id;
-    const gptMessage = await this.getGPTMessage(text, talkerId);
-    const result = `${text}\n ------\n ${gptMessage}`;
-    await this.trySay(room, result);
-  }
-  async onMessage(message: Message) {
-    const talker = message.talker();
-    const rawText = message.text();
-    const room = message.room();
-    const messageType = message.type();
-    const privateChat = !room;
-    if (this.isNonsense(talker, messageType, rawText)) {
-      return;
-    }
-    if (this.tiggerGPTMessage(rawText, privateChat)) {
-      const text = this.cleanMessage(rawText, privateChat);
-      if (privateChat) {
-        return await this.onPrivateMessage(talker, text);
-      } else {
-        return await this.onGroupMessage(talker, text, room);
-      }
-    } else {
-      return;
-    }
   }
 }
